Convert to an Arduino library
The documentation calls this project an "Arduino Library" but, in fact, previously it was an Arduino sketch, not a library. This change makes it a true Arduino library.

I put the modified SoftwareSerial under the utility folder to avoid the IDE displaying 'Multiple libraries were found for "SoftwareSerial.h"' messages for any sketch uses the stock SoftwareSerial library. Although harmless, these messages tend to confuse beginners.

Reference:
https://github.com/arduino/Arduino/wiki/Arduino-IDE-1.5:-Library-specification